Move get_datasets back to the controller to fix unit tests
removes the dependency for ckanapi to run the tests
